Move utility functions from watershed to _util
